Refactor Discord native modal and drop aggregate agent list
