+++ config/urls.py
@@ -217,7 +217,6 @@ from console.discord_api import (
     AgentDiscordChannelsView,
     AgentDiscordConnectView,
     AgentDiscordSubscriptionsView,
-    DiscordAgentConnectionsView,
     DiscordOAuthCallbackView,
     DiscordOAuthStartView,
 )
@@ -596,7 +595,6 @@ urlpatterns = [
     path("console/api/agents/<uuid:agent_id>/discord/connect/", AgentDiscordConnectView.as_view(), name="console-agent-discord-connect"),
     path("console/api/agents/<uuid:agent_id>/discord/guilds/<str:guild_id>/channels/", AgentDiscordChannelsView.as_view(), name="console-agent-discord-channels"),
     path("console/api/agents/<uuid:agent_id>/discord/subscriptions/", AgentDiscordSubscriptionsView.as_view(), name="console-agent-discord-subscriptions"),
-    path("console/api/discord/agents/", DiscordAgentConnectionsView.as_view(), name="console-discord-agent-connections"),
     path("console/api/session/", ConsoleSessionAPIView.as_view(), name="console_session"),
     path("console/api/support/request/", AppSupportRequestAPIView.as_view(), name="console_app_support_request"),
     path("console/api/billing/initial/", BillingInitialDataAPIView.as_view(), name="console_billing_initial_data"),
