+++ frontend/src/components/mcp/PipedreamAppsPanel.tsx
@@ -2,12 +2,13 @@ import { useCallback, useMemo } from 'react'
 import { useQuery } from '@tanstack/react-query'
 import { CheckCircle2, Loader2, Plus, Sparkles } from 'lucide-react'
 
-import { fetchDiscordAgentConnections } from '../../api/discordNative'
+import { fetchAgentRoster } from '../../api/agents'
 import { fetchPipedreamAppSettings, type PipedreamAppSummary } from '../../api/mcp'
 import { fetchNativeIntegrations, type NativeIntegrationProvider } from '../../api/nativeIntegrations'
 import { InlineStatusBanner } from '../common/InlineStatusBanner'
 import { getSettingsSurfaceClassName } from '../common/SettingsSurface'
 import { useModal } from '../../hooks/useModal'
+import { agentHasDiscordNative } from './DiscordNativeAppModal'
 import { withDiscordNativeProviderConnection } from './DiscordNativeShared'
 import { NativeProviderIcon } from './NativeIntegrationShared'
 import { PipedreamAppsModal } from './PipedreamAppsModal'
@@ -44,9 +45,9 @@ export function PipedreamAppsPanel({
     queryFn: () => fetchNativeIntegrations(nativeIntegrationsUrl as string),
     enabled: Boolean(nativeIntegrationsUrl),
   })
-  const discordConnectionsQuery = useQuery({
-    queryKey: ['discord-agent-connections'],
-    queryFn: fetchDiscordAgentConnections,
+  const agentRosterQuery = useQuery({
+    queryKey: ['agent-roster'],
+    queryFn: () => fetchAgentRoster(),
     enabled: Boolean(nativeIntegrationsUrl),
   })
 
@@ -59,10 +60,8 @@ export function PipedreamAppsPanel({
   }), [])
   const effectiveSettings = settingsQuery.data ?? emptySettings
   const discordConnected = useMemo(
-    () => (discordConnectionsQuery.data?.agents ?? []).some((agent) => (
-      agent.connected || agent.subscribed || agent.skillEnabled
-    )),
-    [discordConnectionsQuery.data?.agents],
+    () => (agentRosterQuery.data?.agents ?? []).some(agentHasDiscordNative),
+    [agentRosterQuery.data?.agents],
   )
   const nativeProviders = useMemo(() => {
     return withDiscordNativeProviderConnection(nativeIntegrationsQuery.data?.providers ?? [], discordConnected)
@@ -132,22 +131,22 @@ export function PipedreamAppsPanel({
             type="button"
             className={buttonClassName}
             onClick={openModal}
-            disabled={!canOpenModal || settingsQuery.isLoading || nativeIntegrationsQuery.isLoading || discordConnectionsQuery.isLoading}
+            disabled={!canOpenModal || settingsQuery.isLoading || nativeIntegrationsQuery.isLoading || agentRosterQuery.isLoading}
           >
             <Plus className="h-4 w-4" aria-hidden="true" />
             Add Apps
           </button>
         </div>
 
-        {(hasPipedreamApps && settingsQuery.isLoading) || nativeIntegrationsQuery.isLoading || discordConnectionsQuery.isLoading ? (
+        {(hasPipedreamApps && settingsQuery.isLoading) || nativeIntegrationsQuery.isLoading || agentRosterQuery.isLoading ? (
           <div className={loadingClassName}>
             <Loader2 className="h-4 w-4 animate-spin" />
             Loading apps…
           </div>
-        ) : (hasPipedreamApps && settingsQuery.isError) || nativeIntegrationsQuery.isError || discordConnectionsQuery.isError ? (
+        ) : (hasPipedreamApps && settingsQuery.isError) || nativeIntegrationsQuery.isError || agentRosterQuery.isError ? (
           <div className="px-6 py-5">
             <InlineStatusBanner variant="error" surface={embedded ? 'embedded' : 'standalone'}>
-              {resolvePipedreamAppsErrorMessage(settingsQuery.error ?? nativeIntegrationsQuery.error ?? discordConnectionsQuery.error, 'Unable to load apps right now.')}
+              {resolvePipedreamAppsErrorMessage(settingsQuery.error ?? nativeIntegrationsQuery.error ?? agentRosterQuery.error, 'Unable to load apps right now.')}
             </InlineStatusBanner>
           </div>
         ) : (
